backend/requirements.txt
fastapi==0.85.0
uvicorn==0.17.6
python-multipart==0.0.5
pymongo==4.2.0
Pillow==9.2.0
numpy==1.23.4
python-jose==3.3.0
python-dotenv==0.19.0
